+++ example/lib/views/alerts.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:layrz_icons/layrz_icons.dart';
 import 'package:layrz_theme/layrz_theme.dart';
 import 'package:layrz_theme_example/store/store.dart';
 
@@ -53,6 +54,8 @@ class _AlertsViewState extends State<AlertsView> {
                       type: alertType,
                       size: 40,
                       iconSize: 30,
+                      color: Colors.purple,
+                      icon: LayrzIcons.solarOutlinePlain3,
                     ),
                     const SizedBox(width: 10),
                     for (final style in ThemedAlertStyle.values) ...[
@@ -64,6 +67,8 @@ class _AlertsViewState extends State<AlertsView> {
                             style: style,
                             title: "${alertType.name} - ${style.name}",
                             description: "This is a ${alertType.name} alert with the style ${style.name}.",
+                            color: Colors.purple,
+                            icon: LayrzIcons.solarOutlinePlain3,
                           ),
                         ),
                       ),