+++ lib/src/alerts/src/alert.dart
@@ -16,6 +16,20 @@ class ThemedAlert extends StatelessWidget {
   /// [style] is the style of the alert.
   final ThemedAlertStyle style;
 
+  /// [color] is the color of the alert, used only for [ThemedAlertType.custom].
+  /// If not provided, the default color for the alert type will be used.
+  final Color? color;
+
+  /// [icon] is the icon of the alert, used only for [ThemedAlertType.custom].
+  /// If not provided, the default icon for the alert type will be used.
+  final IconData? icon;
+
+  /// [iconSize] is the size of the icon inside the alert.
+  ///
+  /// By default, on [ThemedAlertStyle.filledIcon] it is set to 25 and on
+  /// [ThemedAlertStyle.layrz] it is set to 22.
+  final double? iconSize;
+
   /// [ThemedAlert] is a widget that displays an alert with a specific type, title, and description.
   ///
   /// Uses internally a [Row] widget to layout the alert content, so, it is required to have dimensions
@@ -27,26 +41,95 @@ class ThemedAlert extends StatelessWidget {
     required this.description,
     this.maxLines = 3,
     this.style = ThemedAlertStyle.layrz,
+    this.color,
+    this.icon,
+    this.iconSize,
   });
 
   @override
   Widget build(BuildContext context) {
+    Color typeColor = type.color ?? Colors.blue;
+    if (type == ThemedAlertType.custom) {
+      typeColor = color ?? type.color ?? Colors.blue;
+    }
+
+    IconData typeIcon = type.icon ?? LayrzIcons.solarOutlineInfoSquare;
+    if (type == ThemedAlertType.custom) {
+      typeIcon = icon ?? type.icon ?? LayrzIcons.solarOutlineInfoSquare;
+    }
+
+    if (style == ThemedAlertStyle.filledIcon) {
+      return Container(
+        decoration: BoxDecoration(
+          color: typeColor,
+          border: Border.all(
+            color: typeColor,
+            width: 2,
+          ),
+          borderRadius: BorderRadius.circular(10),
+        ),
+        clipBehavior: Clip.antiAlias,
+        child: Row(
+          children: [
+            Container(
+              padding: const EdgeInsets.all(10),
+              child: Center(
+                child: Icon(
+                  typeIcon,
+                  color: validateColor(color: typeColor),
+                  size: iconSize ?? 25,
+                ),
+              ),
+            ),
+            Expanded(
+              child: Container(
+                decoration: BoxDecoration(
+                  color: Theme.of(context).scaffoldBackgroundColor,
+                  borderRadius: BorderRadius.only(
+                    topRight: Radius.circular(8),
+                    bottomRight: Radius.circular(8),
+                  ),
+                ),
+                padding: const EdgeInsets.all(10),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      title,
+                      style: context.subtitleStyle,
+                    ),
+                    Text(
+                      description,
+                      style: context.bodyStyle,
+                      maxLines: maxLines,
+                      textAlign: TextAlign.justify,
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ],
+        ),
+      );
+    }
+
     Color? backgroundColor;
     Color? borderColor;
     Color? textColor;
 
     if (style == ThemedAlertStyle.filledTonal) {
-      backgroundColor = type.color.withAlpha((255 * 0.2).toInt());
+      backgroundColor = typeColor.withAlpha((255 * 0.2).toInt());
       borderColor = Colors.transparent;
-      textColor = type.color;
+      textColor = typeColor;
     } else if (style == ThemedAlertStyle.filled) {
-      backgroundColor = type.color;
-      borderColor = type.color;
-      textColor = validateColor(color: type.color);
+      backgroundColor = typeColor;
+      borderColor = typeColor;
+      textColor = validateColor(color: typeColor);
     } else if (style == ThemedAlertStyle.outlined) {
       backgroundColor = Colors.transparent;
-      borderColor = type.color;
-      textColor = type.color;
+      borderColor = typeColor;
+      textColor = typeColor;
     }
 
     bool isLayrzStyle = style == ThemedAlertStyle.layrz;
@@ -71,13 +154,13 @@ class ThemedAlert extends StatelessWidget {
               Container(
                 padding: EdgeInsets.all(isLayrzStyle ? 5 : 0),
                 decoration: BoxDecoration(
-                  color: isLayrzStyle ? type.color.withAlpha((255 * 0.2).toInt()) : null,
+                  color: isLayrzStyle ? typeColor.withAlpha((255 * 0.2).toInt()) : null,
                   borderRadius: BorderRadius.circular(10),
                 ),
                 child: Icon(
-                  type.icon,
-                  color: isLayrzStyle ? type.color : textColor,
-                  size: 20,
+                  typeIcon,
+                  color: isLayrzStyle ? typeColor : textColor,
+                  size: iconSize ?? 22,
                 ),
               ),
               const SizedBox(width: 8),
@@ -94,6 +177,7 @@ class ThemedAlert extends StatelessWidget {
             description,
             style: context.bodyStyle?.copyWith(color: textColor),
             maxLines: maxLines,
+            textAlign: TextAlign.justify,
           ),
         ],
       ),